+++ FlowReports/Properties/Resources.Designer.cs
@@ -232,6 +232,15 @@ namespace FlowReports.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Filter.
+        /// </summary>
+        public static string Filter {
+            get {
+                return ResourceManager.GetString("Filter", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to FlowReport files (*.nas).
         /// </summary>
         public static string FlowReportFiles {